Refactor codegen into lib and bin
